feature: rename files from index
