+++ mobile/src/pages/Cart/index.ts
@@ -1,1 +1,1 @@
-export * from './ui';
+export * from './ui/Cart';
