+++ mobile/src/pages/Catalog/ui/Catalog.tsx
