+++ mobile/src/pages/index.ts
@@ -1,8 +1,8 @@
-export * from './Home/ui';
-export * from './Catalog/ui';
-export * from './Map/ui';
-export * from './Profile/ui';
-export * from './Cart/ui';
+export * from './Home/ui/Home';
+export * from './Catalog/ui/Catalog';
+export * from './Map/ui/Map';
+export * from './Profile/ui/Profile';
+export * from './Cart/ui/Cart';
 export * from './Auth/ui';
-export * from './Settings/ui';
-export * from './Orders/ui';
+export * from './Settings/ui/Settings';
+export * from './Orders/ui/Orders';
