+++ mobile/src/shared/ui/FilterItem/FilterItem.tsx
