+++ mobile/src/shared/ui/ProductItem/ProductItem.tsx
