+++ mobile/src/shared/ui/Promo/Promo.tsx
