+++ mobile/src/shared/ui/index.ts
@@ -1,5 +1,5 @@
-export * from './FilterItem';
-export * from './Input';
-export * from './ProductItem';
-export * from './Promo';
-export * from './CartItem';
+export * from './FilterItem/FilterItem';
+export * from './Input/Input';
+export * from './ProductItem/ProductItem';
+export * from './Promo/Promo';
+export * from './CartItem/CartItem';
